refactor: rename AppNavItem to SidebarNavItem, colocate with AdminSidebar
Only AdminSidebar uses it, and we aren't reworking the top-level
AppMenu in this PR. Moving it next to its sole consumer and giving
it a name that reflects scope.
